Made it easier to define a remote src.
By replacing the host (string) option with an ssh (bool) option you can
instead define the host as part of the src or dest options.

Rathern than actually executing the rsync commands in tests (which
would require an ssh server) the tests just check that the rsync
command looks correct.